requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.18.0
pandas==0.25.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
